Modularize backend + clean uploads tracking
