+++ ProjectSourceCode/src/config/db.js
@@ -1,0 +1,21 @@
+const { Pool } = require('pg');
+
+const pool = new Pool({
+  user: process.env.DB_USER || 'walletwatch_user',
+  host: process.env.DB_HOST || 'localhost',
+  database: process.env.DB_NAME || 'walletwatch',
+  password: process.env.DB_PASS || 'walletwatch_pass',
+  port: Number(process.env.DB_PORT) || 5432
+});
+
+pool
+  .connect()
+  .then(client => {
+    console.log('✅ Database connection successful');
+    client.release();
+  })
+  .catch(error => {
+    console.error('❌ Database connection error:', error.message);
+  });
+
+module.exports = pool;
