+++ ProjectSourceCode/src/middleware/protect.js
@@ -1,0 +1,28 @@
+const jwt = require('jsonwebtoken');
+
+const protect = (req, res, next) => {
+  let token = null;
+
+  if (req.headers.authorization?.startsWith('Bearer')) {
+    token = req.headers.authorization.split(' ')[1];
+  } else if (req.session.token) {
+    token = req.session.token;
+  }
+
+  if (!token) {
+    return res.status(401).json({ message: 'Not authorized, no token' });
+  }
+
+  try {
+    const decoded = jwt.verify(
+      token,
+      process.env.JWT_SECRET || 'supersecretjwt'
+    );
+    req.user = decoded;
+    next();
+  } catch (error) {
+    return res.status(401).json({ message: 'Not authorized, token failed' });
+  }
+};
+
+module.exports = protect;
